requirements.txt
pytest_playwright
otpauth